% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grangersearch-package.R
\docType{package}
\name{grangersearch-package}
\alias{grangersearch}
\alias{grangersearch-package}
\title{grangersearch: Granger Causality Testing for Time Series}
\description{
Performs Granger causality tests on pairs of time series to determine causal relationships. Uses Vector Autoregressive (VAR) models to test whether one time series helps predict another beyond what the series' own past values provide. Returns structured results including p-values, test statistics, and causality conclusions for both directions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nkorf/grangersearch}
  \item Report bugs at \url{https://github.com/nkorf/grangersearch/issues}
}

}
\author{
\strong{Maintainer}: Nikolaos Korfiatis \email{n.korfiatis@uea.ac.uk}

}
\keyword{internal}
