% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grangersearch.R
\name{print.granger_distribution}
\alias{print.granger_distribution}
\title{Print Method for granger_distribution Objects}
\usage{
\method{print}{granger_distribution}(x, ...)
}
\arguments{
\item{x}{A \code{granger_distribution} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for granger_distribution Objects
}
