% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grangersearch.R
\name{print.granger_result}
\alias{print.granger_result}
\title{Print Method for granger_result Objects}
\usage{
\method{print}{granger_result}(x, ...)
}
\arguments{
\item{x}{A \code{granger_result} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for granger_result Objects
}
